requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Computer Vision and Face Recognition
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
numpy==1.24.3
pillow==10.0.1

# Machine Learning and Embeddings
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
faiss-cpu==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.25.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
flake8==6.1.0
